Remove duplicate ConnectedTicket definition
